start migrating to new python tilecutter

+++ src/demos/adventure/scripts/build-tilecut.sh
@@ -16,25 +16,54 @@ function tidyup {
 	exit 1;
 }
 
+if [[ $# != 5 ]]; then 
+	echo "Incorrect number of arguments" 1>2& 
+	exit 1 
+fi
+
+pal="$1"
+charcuts="$2"
+backcuts="$3"
+frontcuts="$4"
+outfile="$5"
+
+if [[ ! -e "$pal" ]]; then 
+	echo "Missing palette file : \"$pal\"" 1>2&
+	exit 1
+fi
+if [[ ! -e "$charcuts" ]]; then
+	echo "Missing charcuts file : \"$charcuts\"" 1>2&
+	exit 1
+fi
+if [[ ! -e "$backcuts" ]]; then 
+	echo "Missing backcuts file : \"$backcuts\"" 1>2&
+	 exit 1
+fi
+if [[ ! -e "$frontcuts" ]]; then
+	echo "Missing frontcuts file : \"$frontcuts\"" 1>2&
+	exit 1
+fi
+
+
 
 echo '<?xml version="1.0" ?>' > "$tmp"
 echo '<tile-cutter>' >> "$tmp"
 
-cat src-graphics/palette.xml >> "$tmp"
+cat "$pal" >> "$tmp"
 
-cat src-graphics/character.cutdefs >> "$tmp"
+cat "$charcuts" >> "$tmp"
 
 echo '<dest file="over-back.til" size-x="16" size-y="24" mask="n">' >> "$tmp"
-echo '<source file="../../src-graphics/overworld16x24.png" />' >> "$tmp"
-cat build/tiles/over-back.cutdefs >> "$tmp"
+echo '<source file="src-graphics/overworld16x24.png" />' >> "$tmp"
+cat "$backcuts" >> "$tmp"
 echo '</dest>' >> "$tmp"
 
 echo '<dest file="over-front.til" size-x="16" size-y="24" mask="y">' >> "$tmp"
-echo '<source file="../../src-graphics/overworld16x24.png" />' >> "$tmp"
-cat build/tiles/over-front.cutdefs >> "$tmp"
+echo '<source file="src-graphics/overworld16x24.png" />' >> "$tmp"
+cat "$frontcuts" >> "$tmp"
 echo '</dest>' >> "$tmp"
 
 
 echo '</tile-cutter>' >> "$tmp"
 
-mv "$tmp" "build/tiles/tile-cutter.xml"
+mv "$tmp" "$outfile"
